feat: tests implementation of mui datatable
