+++ src/components/syncs/SyncDataSet.js
@@ -1,13 +1,23 @@
 import React, { useEffect, useState } from "react";
+import MUIDataTable from "mui-datatables";
 import PortalHeader from "../shared/PortalHeader";
 import PeriodPicker from "../shared/PeriodPicker";
 
 import PluginRegistry from "../core/PluginRegistry";
 
+import {
+  constructDataEntryName,
+  constructDataEntryAssignedTo,
+  constructDataEntryActiveContracts,
+  constructDataEntryActions,
+  constructAccessAndApprovalWorkflow,
+} from "./tables";
+
+import { AccessDisplay } from "./accessDisplay";
+
 import _ from "lodash";
 import {
   Button,
-  Checkbox,
   Typography,
   makeStyles,
   Paper,
@@ -24,27 +34,6 @@ const useStyles = makeStyles({
     verticalAlign: "middle",
   },
 });
-
-const AccessDisplay = ({ access, displayName, dhis2RootUrl }) => {
-  const publicWriteAccess = access.includes("w") && displayName == "Public";
-  const metadataWriteAccess = access[1] == "w";
-  return (
-    <div>
-      <span style={{ fontFamily: "monospace" }}>{access}&nbsp;</span>
-      <span>
-        <a
-          target="_blank"
-          rel="noopener noreferrer"
-          href={dhis2RootUrl + "/dhis-web-maintenance/index.html#/list/dataSetSection/dataSet"}
-        >
-          {displayName}
-        </a>
-        {publicWriteAccess && <span style={{ color: "red" }}> public ! write access !!"</span>}
-        {metadataWriteAccess && <span style={{ color: "red" }}> metadata write access !!"</span>}
-      </span>
-    </div>
-  );
-};
 
 const SyncDataSet = (props) => {
   const dhis2 = PluginRegistry.extension("core.dhis2");
@@ -168,7 +157,7 @@ const SyncDataSet = (props) => {
 
   const addSingleMissingOu = async (myDataSet, missingOrgunits) => {
     setLoading(true);
-    await updateOu(myDataSet, missingOrgunits)
+    await updateOu(myDataSet, missingOrgunits);
     setLoading(false);
     fetchDataSets();
   };
@@ -199,16 +188,46 @@ const SyncDataSet = (props) => {
     fetchDataSets();
   };
 
+  const columns = ["Data Entry", "Assigned To", "Active Contracts", "Actions", "Access & Approval Workflow"];
+  const data = allDataEntries.map((dataEntry) => {
+    // return constructDataForTable(contractsByDataEntryCode, dataEntry, loading, dhis2RootUrl);
+    const dataEntryName = constructDataEntryName(dataEntry);
+    const dataEntryAssignedTo = constructDataEntryAssignedTo(dataEntry);
+    const dataEntryActiveContracts = constructDataEntryActiveContracts(contractsByDataEntryCode, dataEntry);
+    const dataEntryActions = constructDataEntryActions(contractsByDataEntryCode, dataEntry, loading);
+    const dataEntryAccessAndApprovalWorkflow = constructAccessAndApprovalWorkflow(
+      contractsByDataEntryCode,
+      dataEntry,
+      dhis2RootUrl,
+    );
+    return [
+      dataEntryName,
+      dataEntryAssignedTo,
+      dataEntryActiveContracts,
+      dataEntryActions,
+      dataEntryAccessAndApprovalWorkflow,
+    ];
+  });
+
   return (
-    <Paper style={{ minHeight: "85vh", paddingTop: "14px", paddingLeft: "14px"}}>
+    <Paper style={{ minHeight: "85vh", paddingTop: "14px", paddingLeft: "14px" }}>
       <div>
-        <div style={{ display: "flex", flexDirection: "row", alignContent: "center", justifyContent: "flex-start", paddingTop: "5px", paddingLeft:"5px", marginBottom: "20px" }}>
+        <div
+          style={{
+            display: "flex",
+            flexDirection: "row",
+            alignContent: "center",
+            justifyContent: "flex-start",
+            paddingTop: "5px",
+            paddingLeft: "5px",
+            marginBottom: "20px",
+          }}
+        >
           <Typography variant="h6" style={{ marginRight: "20px" }}>
             Dataset Synchronisation for
           </Typography>
           <div style={{ padding: "1px" }}>
             <PeriodPicker
-           
               disableInputLabel={true}
               period={period}
               periodDelta={{
@@ -218,7 +237,7 @@ const SyncDataSet = (props) => {
               onPeriodChange={(newPeriod) => {
                 props.history.push("/sync/datasets/" + newPeriod);
               }}
-             />
+            />
           </div>
         </div>
       </div>
@@ -227,131 +246,7 @@ const SyncDataSet = (props) => {
           Synchronize all {loading && loadingStatus}
         </Button>
       </div>
-      <Table>
-        <TableHead>
-          <TableRow>
-            <th width="10%">Data entry</th>
-            <th width="5%">Assigned to</th>
-            <th width="10%">Active contracts</th>
-            <th width="20%">Actions</th>
-            <th width="40%">Access & approval workflow</th>
-          </TableRow>
-        </TableHead>
-
-        <TableBody>
-          {allDataEntries.map((dataEntry) => {
-            const dataSet = dataSetsById && dataSetsById[dataEntry.dataSetId];
-            const contracts = contractsByDataEntryCode && contractsByDataEntryCode[dataEntry.code];
-            return (
-              <TableRow key={dataEntry.code}>
-                <TableCell>
-                  {dataEntry.name}
-                  <br />
-                  <code>
-                    {dataEntry.code} <br /> {dataEntry.frequency}
-                  </code>
-                </TableCell>
-                <TableCell>
-                  {dataEntry.contracts.map((contract, index) => (
-                    <div>
-                      ( {contract.join(" AND ")} ) {index + 1 < dataEntry.contracts.length && "OR"}
-                    </div>
-                  ))}
-                </TableCell>
-                <TableCell className={classes.aligned}>
-                  {contracts && contracts[0] && contracts[0].activeContracts && contracts[0].activeContracts.length}
-                </TableCell>
-
-                <TableCell>
-                  {!loading &&
-                    contracts &&
-                    contracts.map((contract) => (
-                      <Button
-                        style={{ textAlign: "left" }}
-                        onClick={() => addSingleMissingOu(contract.dataSet, contract.missingOrgunits)}
-                        title={contract.missingOrgunits.map((ou) => ou.name).join(" , ")}
-                        disabled={contract.missingOrgunits.length == 0}
-                      >
-                        Add {contract.missingOrgunits.length} missing OrgUnits to dataset <br />{" "}
-                        {contract.dataSet.name}
-                      </Button>
-                    ))}
-
-                  {!loading && dataEntry.hesabuInputs && contracts && contracts[0] && (
-                    <span
-                      title={
-                        contracts &&
-                        "missing : " +
-                          contracts[0].missingDataElements
-                            .map((de) => (dataElementsById[de] ? dataElementsById[de].name : de))
-                            .join(" , ") +
-                          "\n\n based on : \n" +
-                          dataEntry.hesabuInputs.join(" , ") +
-                          "\n\n is supposed to contain : \n" +
-                          contracts[0].expectedDataElements
-                            .map((de) => (dataElementsById[de] ? dataElementsById[de].name : de))
-                            .join(" ,  )")
-                      }
-                    >
-                      <Button
-                        onClick={() => addMissingDe(dataEntry)}
-                        disabled={contracts && contracts[0].missingDataElements.length == 0}
-                      >
-                        Add {contracts && contracts[0].missingDataElements.length} DataElements to dataset
-                      </Button>
-                    </span>
-                  )}
-                  {dataEntry.hesabuPackage && dataEntry.hesabuInputs == undefined && (
-                    <div style={{ color: "red" }}>No hesabuInputs configured in dataentries.json</div>
-                  )}
-                </TableCell>
-                <TableCell>
-                  <div style={{ display: "block" }}>
-                    {contracts &&
-                      contracts.map((contract) => {
-                        return (
-                          <>
-                            <b style={{ color: "grey" }}>{contract.dataSet.name}</b>{" "}
-                            <a
-                              href={
-                                dhis2RootUrl +
-                                "/dhis-web-maintenance/index.html#/edit/dataSetSection/dataSet/" +
-                                contract.dataSet.id
-                              }
-                              target="_blank"
-                              rel="noopener noreferrer"
-                            >
-                              <code>{contract.dataSet.id}</code>
-                            </a>{" "}
-                            <br />
-                            {contract.dataSet.workflow && "Data approval : " + contract.dataSet.workflow.name+" "+contract.dataSet.workflow.periodType}
-                            {contract.dataSet.workflow == undefined && (
-                              <span style={{ color: "red" }}>no data approval configured</span>
-                            )}
-                            <AccessDisplay
-                              access={contract.dataSet.publicAccess}
-                              displayName={"Public"}
-                              dhis2RootUrl={dhis2RootUrl}
-                            />
-                            {contract.dataSet.userGroupAccesses.map((uga) => (
-                              <AccessDisplay
-                                key={uga.displayName}
-                                access={uga.access}
-                                displayName={uga.displayName}
-                                dhis2RootUrl={dhis2RootUrl}
-                              />
-                            ))}
-                            <br />
-                          </>
-                        );
-                      })}
-                  </div>
-                </TableCell>
-              </TableRow>
-            );
-          })}
-        </TableBody>
-      </Table>
+        <MUIDataTable title={"testing"} data={data} columns={columns} />
     </Paper>
   );
 };
